requirements.txt
Django==2.1.4
wagtail==2.4
wagtailmenus==2.12
wagtail-condensedinlinepanel==0.5.2
django-webpack-loader==0.6.0
django-model-utils==3.1.2

gunicorn==19.9.0
sentry_sdk==0.10.1

psycopg2 == 2.8.3 ; sys_platform != 'win32'
psycopg2-binary == 2.8.3 ; sys_platform == 'win32'
